// apps/fx/src/models/clients.ts
export interface IClients {
  uuid: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
  username: string;
  firstname: string;
  lastname: string;
  hasAcceptedTerms: boolean;
  archivedAt?: string;
  invitedBy?: string;
  verificationCode?: string;
  verifiedAt?: string;
  role?: string;
  twoFASecret?: string;
  twoFA?: string;
  clients?: Array<IUserClient>;
  isMobile2FAEnabled: boolean;
  contact: IClientContact;
}

export interface ISubAccount {
  accountHolderId: string;
  clientType: string;
  friendlyName: string;
  taken: boolean;
}

export interface IUserClient {
  uuid: string;
  createdAt: string;
  deletedAt: string;
  name: string;
  updatedAt: string;
  account: IClientAccount;
  users: Array<IClients>;
  client?: string;
}
// ...
export interface IClientUser {
  uuid: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string;
  username: string;
  firstname: string;
  lastname: string;
  hasAcceptedTerms: boolean;
  archivedAt?: string;
  invitedBy: string;
  verificationCode: string;
  verifiedAt?: string;
  role: string;
  twoFASecret?: string;
  twoFA?: string;
  isMobile2FAEnabled?: boolean;
  documents: Array<any>;
}

export interface IBankMetadata {
  bankName: string;
  branch: string;
  bankCountry: string;
  accountHolderName: string;
  sortCode: string;
  accountNumber: string;
  IBAN: string;
  bicSwift: string;
  currency: string;
  sort1: string;
  uuid: string;
}

export interface IClientContact {
  uuid: string;
  user: IClientUser;
  createdAt: string;
  updatedAt: string;
  businessRole: string;
  country: string;
  mobileNumber: string;
  isApproved: boolean;
  isSubAccount: boolean;
  account: IClientAccount;
  deletedAt?: string;
  cloudCurrencyId?: string;
  openPaydId?: string;
  expectedVolumeOfTransactions?: string;
  expectedValueOfTurnover?: string;
  complyLaunchId?: string;
  complyLaunchResponse?: string;
  currencyCloudPasswordUrl?: string;
  bankMetadata?: IBankMetadata;
  individualMetadata?: IIndividualMetadata;
  invitation?: string;
}

export interface IIndividualMetadata {
  uuid: string;
  createdAt: Date;
  updatedAt: Date;
  title: string;
  firstname: string;
  lastname: string;
  formerName: string;
  otherName: string;
  dateOfBirth: Date;
  placeOfBirth: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  postcode: string;
  state: string;
  country: string;
  emailVerified: string;
  previousAddressLine1: string;
  previousAddressLine2: string;
  previousCity: string;
  previousPostcode: string;
  previousCountry: string;
  previousState: string;
  nationality: string;
  gender: string;
  email: string;
  identificationNumber: string;
  identificationType: string;
  occupation: string;
  employerName: string;
  employerAddress1: string;
  employerAddress2: string;
  employerAddress3: string;
  publicPosition: string;
  highProfilePosition: string;
  deletedAt?: Date;
}
// ...
export interface IClientAccount {
  uuid: string;
  createdAt: string;
  updatedAt: string;
  entityType: string;
  businessMetadata: IClientBusiness;
  riskAssessments: Array<IRiskAssessment>;
  contact: Array<IClientContact>;
  brokers: Array<IBrokers>;
  directors: Array<IClientDirectors>;
  shareholders: Array<IShareholder>;
  isApproved: false;
  pendingMetadatas: Array<any>;
  deletedAt?: string;
  bankMetadata?: IBankMetadata;
  individualMetadata?: IIndividualMetadata;
  cloudCurrencyId?: string;
  openPaydId?: string;
  archivedAt?: string;
  gatewayId?: string;
  gateway?: string;
  fee?: string;
}